refactor(promise): split the new exception helpers into promise_errors.rs
The previous commit's additions pushed `methods_promise.rs` from 476 to 557
lines, over the project's 500-line-per-file limit. Move the helpers it added --
the `X::Promise::Vowed` / `X::Promise::Resolved` builders, the mixin-transparent
`is_exception_object` / `exception_instance_of` shape checks, and the
`X::Promise::Broken` composition plus its throw-site backtrace stamp -- into a
new `runtime/promise_errors.rs`, next to the `promise_broken_gist.rs` that
renders that role's gist.

Pure code motion: `methods_promise.rs` is back to 457 lines and no behaviour
changes. Verified with the 41-assertion
`t/promise-keep-break-semantics.t`, the 126-file
`t/promise*`/`supply*`/`react*`/`whenever*`/`lock*` set, and 873 Rust unit
tests.
