+++ src/runtime/mod.rs
@@ -538,6 +538,7 @@ mod ops_set;
 mod output_sink;
 pub(crate) mod phasers;
 mod promise_broken_gist;
+mod promise_errors;
 mod react_died;
 pub(crate) mod react_done_handler_depth;
 mod receiver_class;
